畫布無法先顯示
